CleanUp-Code
Created google-auth.js to make code cleaner
